+++ src/train/lyric_generation/model_creator.py
@@ -9,8 +9,7 @@ from tensorflow.keras.metrics import CategoricalAccuracy
 from tensorflow.keras.optimizers import RMSprop
 
 from constants import Constants
-from train.lyric_generation.data_loader import DataContainer
-from train.utils import Utils
+from train.utils import DataContainer, Utils
 
 
 class ModelCreator:
